A little more work on HVAC.
Ensure that HVAC monitoring is on in lutron.go

Add parsing and handle skeletons in hvac.go
The big question is, given that there are so many
channels of data, how do we handle readers, and
how do we handle monitors?  Is there a reader queue
for each channel?  What type do we pass to monitors
to say what has changed/what values there are?
Should we try to consolidate messages for monitors
(e.g., when you hit eco, you get "echo on" plus
"setpoints have changed" messages, should that be
one monitor message?)